# api/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Perpustakaan",
    "description": "Perpustakaan API Collection",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3880/"
    }
  ],
  "paths": {
...
    "/menu": {
      "post": {
        "tags": [
          "Menu"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LevelAdd"
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Menu"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LevelUpdate"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Menu"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LevelDelete"
              }
            }
          }
        }
      }
    },
...
    "/menu/{id}": {
      "get": {
        "tags": [
          "Menu"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsersAdd"
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsersUpdate"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsersDelete"
              }
            }
          }
        }
      }
    },
...
    "/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LevelAdd": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "example": "level label"
          }
        }
      },
      "LevelUpdate": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "example": "level label"
          },
          "id": {
            "type": "number",
            "example": 0
          }
        }
      },
      "LevelDelete": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 0
          }
        }
      },
...
      "UsersAdd": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "example": "users label"
          }
        }
      },
      "UsersUpdate": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "example": "users label"
          },
          "id": {
            "type": "number",
            "example": 0
          }
        }
      },
      "UsersDelete": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 0
          }
        }
      },
...
    }
  }
}